requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.33
pytest==7.4.0
